requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
pyyaml
pytest
httpx
opencv-python
pydantic